Replace path.ts with types from @bufbuild/protobuf/reflect
