+++ packages/protovalidate/src/error.ts
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import { type Path, pathToString } from "./path.js";
+import { pathToString, type Path } from "@bufbuild/protobuf/reflect";
 import {
   type FieldPath,
   type FieldPathElement,
